+++ src/window.rs
@@ -148,6 +148,34 @@ impl<'w> Window<'w> {
 		}
 	}
 
+	pub fn get_character(&self) -> Input {
+		loop {
+			let c = loop {
+				let c = self.window.getch();
+				if c.is_some() {
+					break c.unwrap();
+				}
+			};
+
+			match c {
+				PancursesInput::Character(c) if c == '\n' => break Input::Enter,
+				PancursesInput::Character(c) => break Input::Character(c),
+				PancursesInput::KeyEnter => break Input::Enter,
+				PancursesInput::KeyBackspace => break Input::Backspace,
+				PancursesInput::KeyDC => break Input::Delete,
+				PancursesInput::KeyRight => break Input::MoveCursorRight,
+				PancursesInput::KeyLeft => break Input::MoveCursorLeft,
+				PancursesInput::KeyResize => {
+					pancurses::resize_term(0, 0);
+					self.height.replace(self.window.get_max_y());
+					self.width.replace(self.window.get_max_x());
+					break Input::Resize
+				},
+				_ => {}
+			};
+		}
+	}
+
 	pub fn get_input(&self) -> Input {
 		// ignore None's, since they are not really valid input
 		let c = loop {
@@ -164,13 +192,14 @@ impl<'w> Window<'w> {
 			PancursesInput::Character(c) if c == 'Q' => Input::ForceAbort,
 			PancursesInput::Character(c) if c == 'w' => Input::Rebase,
 			PancursesInput::Character(c) if c == 'W' => Input::ForceRebase,
-			PancursesInput::Character(c) if c == 'p' => Input::Pick,
-			PancursesInput::Character(c) if c == 'b' => Input::Break,
-			PancursesInput::Character(c) if c == 'r' => Input::Reword,
-			PancursesInput::Character(c) if c == 'e' => Input::Edit,
-			PancursesInput::Character(c) if c == 's' => Input::Squash,
-			PancursesInput::Character(c) if c == 'f' => Input::Fixup,
-			PancursesInput::Character(c) if c == 'd' => Input::Drop,
+			PancursesInput::Character(c) if c == 'p' => Input::ActionPick,
+			PancursesInput::Character(c) if c == 'b' => Input::ActionBreak,
+			PancursesInput::Character(c) if c == 'r' => Input::ActionReword,
+			PancursesInput::Character(c) if c == 'e' => Input::ActionEdit,
+			PancursesInput::Character(c) if c == 's' => Input::ActionSquash,
+			PancursesInput::Character(c) if c == 'f' => Input::ActionFixup,
+			PancursesInput::Character(c) if c == 'd' => Input::ActionDrop,
+			PancursesInput::Character(c) if c == 'E' => Input::Edit,
 			PancursesInput::Character(c) if c == 'j' => Input::SwapSelectedDown,
 			PancursesInput::Character(c) if c == 'k' => Input::SwapSelectedUp,
 			PancursesInput::KeyDown => Input::MoveCursorDown,